requirements.txt
# -e git://github.com/onepercentclub/bluebottle.git@v0.0.16#egg=bluebottle
-e git://github.com/onepercentclub/bluebottle.git@meta-seo-2#egg=bluebottle


Django==1.5.4
Pillow==2.1.0

django-localflavor==1.0

django_extensions==1.1.1

django_debug_toolbar

raven # Sentry client (remote logging)
django-celery==3.0.17 # Asynchronous task server

# Django Nose integration (for test suites)
# Nose is not the default testrunner due to
# https://github.com/jbalogh/django-nose/issues/85
django-nose

# CSS/JS compression - http://pypi.python.org/pypi/django_compressor/1.2
cssmin==0.1.4
jsmin==2.0.2
BeautifulSoup==3.2.1
django_compressor==1.2

# pytz is "highly recommended" by the Django timezone documentation since
# we're using Django timezone support
# https://docs.djangoproject.com/en/1.4/topics/i18n/timezones/
pytz

sphinx # Sphinx documentation

# http://pypi.python.org/pypi/sorl-thumbnail/11.12
sorl-thumbnail==11.12

# http://pypi.python.org/pypi/django-choices/
django-choices==1.1.11

# http://pypi.python.org/pypi/surlex/0.1.2
surlex==0.1.2

# http://pypi.python.org/pypi/South
south==0.8.1

# Django-REST-Framework
django-filter==0.6
djangorestframework==2.3.6

# Jenkins integration
django-jenkins==0.14.0

# Coverage in tests
coverage==3.6

# Tagging
django-taggit==0.10a1
django-taggit-autocomplete-modified==0.1.0b4

# Legacy authentication module for authenticating users from the old PHP site.
# https://github.com/onepercentclub/legacyauth/archive/3f2406c50dead25a748fb2433de55b73a9162f18.zip
-e git://github.com/onepercentclub/legacyauth.git@3f2406c50dead25a748fb2433de55b73a9162f18#egg=legacyauth

# Gunicorn WSGI server
gunicorn

# Embedding videos and other media
# https://github.com/coleifer/micawber
micawber==0.2.6

# Used for WallPosts and Blog/News content (as a dependency of django-fluent-contents).
django_polymorphic==0.5

# Blog/news contents is created using blocks
django-fluent-contents==0.9a1
django-wysiwyg==0.5.1
django-tinymce==1.5.1b2
html5lib==0.95

# Admin dashboard improvements
django-fluent-dashboard==0.3.2
django-admin-tools==0.5.1

# Fabric deployment helper
# WARNING: Make sure it's installed in your environment and not system-wide!
#          The Fabric script depends on gitpython which needs to be available.
fabric
gitpython

# Transifex - for translations!
transifex-client==0.9

# Handlebars in templates
django-templatetag-handlebars==1.2.0

# Extend AND override app templates
django-apptemplates==0.0.1

# Cowry dependencies.
django_countries==1.5
requests==1.2.3
# This version should be considered suds-jurko 0.4.1 jurko pre 5.
# This is just a fork of the official repo in case it disappears.
https://bitbucket.org/onepercentclub/suds/get/afe727f50704.zip

django-iban==0.2.1

django-social-auth==0.7.23

# SalesForce integration.
# django-salesforce==0.1.6.2
# setup.py doesn't seem to be working properly so we're using a fixed version.
# https://github.com/freelancersunion/django-salesforce/issues/4
# https://github.com/onepercentclub/django-salesforce/archive/1e54beb7bcc15a893e9590fb27cbf08853da5599.zip
-e git://github.com/onepercentclub/django-salesforce.git@1e54beb7bcc15a893e9590fb27cbf08853da5599#egg=django-salesforce

# There hasn't been a release of django-registration is a while and we need the 1.5 custom user support.
# Official: https://bitbucket.org/ubernostrum/django-registration/
# Fork with 1.5 Custom User support: https://bitbucket.org/eire1130/django-registration
# Our fork of the fork with HTML_ACTIVATION_EMAIL option:
# TODO Try to unwind this mess and contribute some stuff upstream.
https://bitbucket.org/onepercentclub/django-registration/get/ae9e9ed265ed.zip


# Private uploads
https://bitbucket.org/wkornewald/django-filetransfers/get/32ddeac.zip

# DKIM to sign e-mails
dkimpy
dnspython

# Functional testing libraries
selenium==2.33.0
splinter==0.5.3

# Mock library for testing
mock==1.0.1

# JS translation with static files
django-statici18n==0.4.5

# Communication with the old DocData account.
# https://github.com/onepercentclub/django-docdata/archive/120ae5b8a1da6152d43d4601edc8832268e05515.zip
-e git://github.com/onepercentclub/django-docdata.git@120ae5b8a1da6152d43d4601edc8832268e05515#egg=django-docdata
pycurl==7.19.0

# Date and Currency formatting.
babel==1.3
